requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv